InterviewAI/requirements.txt
streamlit==1.28.1
google-generativeai==0.3.2
PyPDF2==3.0.1
python-docx==0.8.11
python-multipart==0.0.6
pandas==2.1.3
numpy==1.24.3
plotly==5.17.0
reportlab==4.0.7
fpdf2==2.7.6
python-dotenv==1.0.0
openai-whisper==20231117
soundfile==0.12.1
streamlit-audio-recorder==0.0.8
pyaudio==0.2.11
